added the checkIfMobileScreen as a function and imported it into the different components, also adjusted the file structure and namings to be more coherent
