+++ src/checkIfMobileScreen.js
@@ -1,0 +1,17 @@
+import { useState, useEffect } from "react";
+
+export default function checkIfMobileScreen() {
+  const [width, setWidth] = useState(window.innerWidth);
+  const handleWindowSizeChange = () => {
+    setWidth(window.innerWidth);
+  };
+
+  useEffect(() => {
+    window.addEventListener("resize", handleWindowSizeChange);
+    return () => {
+      window.removeEventListener("resize", handleWindowSizeChange);
+    };
+  }, []);
+
+  return width <= 768;
+}
