+++ src/components/Filters/index.jsx
@@ -1,4 +1,4 @@
-import { useState, useContext } from "react";
+import { useContext } from "react";
 import styles from "./Filters.module.css";
 import { FilterContext } from "../ProductCards";
