+++ src/components/Layout/Header/index.jsx
@@ -1,6 +1,5 @@
 import { Link } from "react-router-dom";
 import LogoIcon from "../../../assets/logos/logo.png";
-import BackArrowIcon from "../../../assets/icons/back.svg";
 import Searchbar from "./Search";
 import HamburgerMenu from "./Nav";
 import Cart from "./Cart";
@@ -18,9 +17,6 @@ export default function Header({ cart }) {
         <Link to="/" className={styles.icon} onClick={() => handleClick()}>
           <img src={LogoIcon} alt="Lazz logo, click to go to home page" />
         </Link>
-        {/* <div className={styled.icon}>
-        <img src={BackArrowIcon} alt="fill-in-later" />
-      </div> */}
         <Searchbar />
         <Link to="checkout" className={styles.icon}>
           <Cart cart={cart} />
