+++ src/components/pages/Checkout/Checkout.module.css
@@ -1,4 +1,4 @@
-.wrapper {
+.wrapper2 {
   margin-top: 90px;
   margin-bottom: 150px;
   padding: 1rem;
@@ -17,4 +17,70 @@
   @media screen and (min-width: 1200px) {
     padding: 0rem;
   }
+}
+
+.wrapper {
+  width: 100%;
+  display: flex;
+  flex-direction: column;
+  gap: 1rem;
+
+  @media screen and (min-width: 768px) {
+    margin-bottom: 2rem;
+  }
+}
+
+.cardWrapper {
+  display: flex;
+  gap: 0.5rem;
+  background-color: var(--color-white);
+  height: 125px;
+}
+
+.imageWrapper {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  img {
+    width: 100px;
+    object-fit: cover;
+  }
+}
+
+.productInfo {
+  display: flex;
+  flex-direction: column;
+  /* justify-content: space-between; */
+  width: 100%;
+  height: 100%;
+  /* gap: 0.3rem; */
+  justify-content: space-around;
+}
+
+.infoWrapper {
+  padding: 1rem;
+  width: 100%;
+  display: flex;
+  flex-direction: column;
+  justify-content: space-between;
+  height: 100%;
+
+
+  h2 {
+    font-size: 1.125rem;
+    font-weight: 600;
+    color: var(--color-black);
+  }
+}
+
+.container {
+  display: flex;
+  justify-content: space-between;
+  align-items: end;
+  /* height: 100%; */
+}
+
+.quantityWrapper {
+  justify-self: end;
+  align-self: flex-end;
 }
