+++ src/components/pages/Home/index.jsx
@@ -1,7 +1,6 @@
 import { ProductsToDisplay } from "../../ProductCards";
-import Categories from "./Categories";
-import BrandCommercial from "./BrandCommercial";
-import Filters from "../../Filters";
+import Categories from "../../Categories";
+import BrandCommercial from "../../BrandCommercial";
 import styles from "./Home.module.css";
 
 export default function Home() {
