+++ src/components/quanity/index.jsx
@@ -1,6 +1,5 @@
 import styles from "./Quantity.module.css";
 import { useReducer, createContext, useEffect } from "react";
-import { useContext } from "react";
 import { useOutletContext } from "react-router-dom";
 
 function reducer(state, action) {
@@ -48,7 +47,7 @@ function reducer(state, action) {
 export const QuantityContext = createContext();
 
 export function Quantity({ page, product }) {
-  const { cart, setCart, totalSum, setTotalSum } = useOutletContext();
+  const { setCart, totalSum, setTotalSum } = useOutletContext();
   const initialState = { product: product, total: totalSum, count: product.quantity };
   const [state, dispatch] = useReducer(reducer, initialState);
